commit Validaciones

+++ ludikApi/LogicaAplicacion/ImplementacionCasosUsos/TablaClasificacion/BajaTablaClasificacion.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using InterfacesRepositorio;
 using LogicaAplicacion.InterfacesCasosUsos.TablaClasificacion;
+using LogicaNegocio.Entidades;
 using LogicaNegocio.Resultados;
 
 namespace LogicaAplicacion.ImplementacionCasosUsos.TablaClasificacion
@@ -12,18 +13,35 @@ namespace LogicaAplicacion.ImplementacionCasosUsos.TablaClasificacion
     public class BajaTablaClasificacion:IBajaTablaClasificacion
     {
         private readonly IRepositorioTablasClasificacion _repo;
+        private readonly IRepositorioProfesores _repoProfesores;
 
-        public BajaTablaClasificacion(IRepositorioTablasClasificacion repo)
+
+        public BajaTablaClasificacion(IRepositorioTablasClasificacion repo,IRepositorioProfesores repositorioProfesores)
         {
             _repo = repo;
+            _repoProfesores = repositorioProfesores;
         }
 
-        public async Task<Resultado> EjecutarAsync(int tablaId)
+        public async Task<Resultado> EjecutarAsync(int tablaId, string profesorId)
         {
-            //tiene que validar que el grupo de la tabla de clasificacion , su profesor Id es igual al profesor logueado
+            var resBusqueda = await _repo.GetByIdAsync(tablaId);
+            var tabla = resBusqueda.Valor;
+            var resBusquedaProfesor = await _repoProfesores.GetByStringIdAsync(profesorId);
+            var profesor = resBusquedaProfesor.Valor;
+
+            bool tieneEseGrupo = profesor.Grupos
+            .Any(g => g.Id == tabla.GrupoId);
+
+            if (!tieneEseGrupo)
+            {
+                return Resultado.Falla(new Error(
+                    "Error.Autorizacion",
+                    "El grupo de la tabla no pertenece al profesor logueado."));
+            }
             var res = await _repo.RemoveAsync(tablaId);
             if (res.EsFallo)
                 return Resultado.Falla(res.Errores);
+            
 
             return Resultado.Exitoso();
         }
